+++ backend/src/SkillCraft.Tools.Infrastructure.PostgreSQL/Migrations/20250116025836_Release_0_5_0.Designer.cs
@@ -2,6 +2,7 @@
 using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 using SkillCraft.Tools.Infrastructure;
@@ -11,9 +12,11 @@ using SkillCraft.Tools.Infrastructure;
 namespace SkillCraft.Tools.Infrastructure.PostgreSQL.Migrations
 {
     [DbContext(typeof(SkillCraftContext))]
-    partial class SkillCraftContextModelSnapshot : ModelSnapshot
+    [Migration("20250116025836_Release_0_5_0")]
+    partial class Release_0_5_0
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder