+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-Nothing yet.
+### Fixed
+
+- Added missing migration.
 
 ## [0.5.0] - 2025-01-15